backend/requirements.txt
beautifulsoup4>=4.12
requests>=2.31
supabase>=2.0
python-dotenv>=1.0
textblob>=0.18
schedule>=1.2
lxml>=5.1
